backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
ultralytics>=8.2.0
opencv-python-headless>=4.9.0
python-multipart==0.0.9
firebase-admin>=6.5.0
python-dotenv==1.0.1
pillow>=10.0.0
numpy>=1.26.0
aiofiles==23.2.1
